Created new component - MessagesList; for list of messages
